Remove sending functional from `Request`
The request should be a simple structure that we'll expose in
exceptions. It don't know how to send itself to API, only how to
build itself (uri, and params) and (future features) how to
validate itself against API specification.

The `#call` method and handling results is removed to
separate class (namely, `Sender`).